additional refactoring after merge with master
