+++ spring-cloud-gcp-data-spanner/src/test/java/org/springframework/cloud/gcp/data/spanner/core/convert/MappingSpannerReadConverterTest.java
@@ -26,9 +26,9 @@ import com.google.common.collect.ImmutableList;
 import org.junit.Before;
 import org.junit.Test;
 
-import org.springframework.cloud.gcp.data.spanner.core.convert.Entities.FaultyTestEntity;
-import org.springframework.cloud.gcp.data.spanner.core.convert.Entities.OuterTestEntity;
-import org.springframework.cloud.gcp.data.spanner.core.convert.Entities.TestEntity;
+import org.springframework.cloud.gcp.data.spanner.core.convert.TestEntities.FaultyTestEntity;
+import org.springframework.cloud.gcp.data.spanner.core.convert.TestEntities.OuterTestEntity;
+import org.springframework.cloud.gcp.data.spanner.core.convert.TestEntities.TestEntity;
 import org.springframework.cloud.gcp.data.spanner.core.mapping.SpannerDataException;
 import org.springframework.cloud.gcp.data.spanner.core.mapping.SpannerMappingContext;
 import org.springframework.core.convert.ConversionFailedException;
